Rewrite favourite, recent and search screens using reusable table and collection
